requirements.txt
python-dotenv==1.0.0
python-telegram-bot[job-queue]==20.4
google-api-python-client==2.95.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
pyyaml==6.0
pytz==2024.1
python-dateutil==2.9.0
